+++ changelog.md
@@ -17,6 +17,17 @@ Upgraded software (GNOME):
 
 - gnome-software: `42.3 --> 42.4`
 - libadwaita: `1.1.3 --> 1.1.4`
+- eog: `42.2 --> 42.3`
+- geocode-glib: `3.26.3 --> 3.26.4`
+- gtk4: `4.6.6 --> 4.7.1`
+- gjs: `1.73.1 --> 1.73.2`
+- gnome-desktop: `42.3 --> 42.4`
+- gssdp: `1.5.1 --> 1.5.2`
+- gupnp: `1.5.2 --> 1.5.3`
+- gtksourceview5: `5.5.0 --> 5.5.1`
+- rygel: `0.41.1 --> 0.41.2`
+
+
 
 # MassOS 2022.08
 First official release of MassOS with GNOME support.